VYBES-109: enhance project organization
