+++ app/_layout.tsx
@@ -35,6 +35,7 @@ export default function RootLayout() {
         <Stack.Screen name='login' options={{headerShown: false}} />
         <Stack.Screen name='debug' options={{headerShown: false}} />
         <Stack.Screen name='(tabs)' options={{headerShown: false}} />
+        <Stack.Screen name='playlist' options={{headerShown: false}} />
         <Stack.Screen name='+not-found' />
       </Stack>
       <StatusBar style='auto' />
