+++ app/playlist/_layout.tsx
@@ -1,0 +1,20 @@
+import {Stack} from 'expo-router';
+
+export default function PlaylistLayout() {
+  return (
+    <Stack>
+      <Stack.Screen
+        name='[id]'
+        options={{
+          headerShown: false,
+        }}
+      />
+      <Stack.Screen
+        name='track/[id]'
+        options={{
+          headerShown: false,
+        }}
+      />
+    </Stack>
+  );
+}
