+++ app/playlist/track/_layout.tsx
@@ -1,0 +1,14 @@
+import {Stack} from 'expo-router';
+
+export default function TrackLayout() {
+  return (
+    <Stack>
+      <Stack.Screen
+        name='[id]'
+        options={{
+          headerShown: false,
+        }}
+      />
+    </Stack>
+  );
+}
